@quasiben 2020-07-22
That is so cool!  Thanks @jakirkham !  And again, thank you for providing timing reports as well.